+++ src/gql/types/generators/EntryType.php
@@ -20,6 +20,7 @@ use craft\gql\types\elements\Entry;
 use craft\helpers\Gql;
 use craft\helpers\Gql as GqlHelper;
 use craft\models\EntryType as EntryTypeModel;
+use craft\models\GqlSchema;
 
 /**
  * Class EntryType
@@ -34,16 +35,11 @@ class EntryType extends Generator implements GeneratorInterface, SingleGenerator
      */
     public static function generateTypes($context = null): array
     {
-        $entryTypes = Craft::$app->getSections()->getAllEntryTypes();
         $gqlTypes = [];
 
+        $entryTypes = GqlHelper::getSchemaContainedEntryTypes();
+
         foreach ($entryTypes as $entryType) {
-            $requiredContexts = EntryElement::gqlScopesByContext($entryType);
-
-            if (!GqlHelper::isSchemaAwareOf($requiredContexts)) {
-                continue;
-            }
-
             // Generate a type for each entry type
             $type = static::generateType($entryType);
             $gqlTypes[$type->name] = $type;